Decouple wire-representation from struct
Exposing both flags and fields introduces the problem of not having a single
source of truth. For example, when we'll serialized, what if the users passes a
Connect packet with a password but with password_flag == false? This should
remove the ambiguity by leaving a single source of truth for all info in the structs.
Doing this also removes the payload layer in the Packet, since it's not needed anymore.

Signed-off-by: Riccardo Binetti <rbino@gmx.com>